requirements.txt
mqt.qao==0.3.0
qiskit==1.3.2
qiskit-algorithms==0.3.0
networkx==3.5
matplotlib==3.10.6
numpy==2.3.2
sympy==1.14.0